+++ src/main/java/domain/rule/BlackjackRule.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.rule;
 
+import domain.GameResult;
 import domain.card.Cards;
 import java.util.Set;
