+++ src/test/java/domain/BlackjackRuleTest.java
@@ -6,6 +6,8 @@ import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
 import domain.card.Cards;
+import domain.rule.BlackjackRule;
+import domain.rule.GameRule;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
